search/bleve: add SortOrder to RadioQuery that returns a useable search.SortOrder
